requirements.txt
# Core ML/DL
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1
scikit-image>=0.21.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Optimization
tensorflow-model-optimization>=0.7.5
tflite-support>=0.4.4

# API & Server
Flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
aiohttp>=3.8.0
websockets>=11.0

# Database & Storage
firebase-admin>=6.2.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0
loguru>=0.7.0

# Raspberry Pi specific (installer séparément)
# picamera2
# RPi.GPIO
# smbus2